Converted PlayerDataStore to an abstract class, extracting common functionality.
