+++ pom.xml
@@ -53,6 +53,13 @@
       <version>2.11.5</version>
       <scope>provided</scope>
     </dependency>
+
+    <!-- MySQL -->
+    <dependency>
+      <groupId>mysql</groupId>
+      <artifactId>mysql-connector-java</artifactId>
+      <version>8.0.33</version>
+    </dependency>
   </dependencies>
 
   <build>
